+++ 5.2/5.2.0/python2/docker-entrypoint.sh
@@ -1,16 +1,10 @@
 #!/bin/bash
 set -e
 
-COMMANDS="debug help logtail show stop adduser fg kill quit run wait console foreground logreopen reload shell status"
-START="start restart zeoserver"
-CMD="bin/instance"
-LOGFILE="/data/log/instance.log"
+COMMANDS="adduser debug fg foreground help kill logreopen logtail reopen_transcript run show status stop wait"
+START="console start restart"
 
 gosu plone python /docker-initialize.py
-
-# XXX To be removed within Plone 5.2rc2+
-# See https://github.com/plone/Installers-UnifiedInstaller/issues/86
-sed -i "s|Products.DocFinderTab||g" develop.cfg
 
 if [ -e "custom.cfg" ]; then
   if [ ! -e "bin/develop" ]; then
@@ -20,45 +14,20 @@ if [ -e "custom.cfg" ]; then
   fi
 fi
 
+# ZEO Server
 if [[ "$1" == "zeo"* ]]; then
-  CMD="bin/$1"
-  LOGFILE="/data/zeoserver/zeoserver.log"
+  exec gosu plone bin/$1 fg
 fi
 
-if [ -z "$HEALTH_CHECK_TIMEOUT" ]; then
-  HEALTH_CHECK_TIMEOUT=1
-fi
-
-if [ -z "$HEALTH_CHECK_INTERVAL" ]; then
-  HEALTH_CHECK_INTERVAL=1
-fi
-
+# Plone instance start
 if [[ $START == *"$1"* ]]; then
-  _stop() {
-    gosu plone $CMD stop
-    kill -TERM $child 2>/dev/null
-  }
-
-  trap _stop SIGTERM SIGINT
-  gosu plone touch $LOGFILE
-  gosu plone $CMD start
-  gosu plone tail -f $LOGFILE &
-  child=$!
-
-  pid=`$CMD status | sed 's/[^0-9]*//g'`
-  if [ ! -z "$pid" ]; then
-    echo "Application running on pid=$pid"
-    sleep "$HEALTH_CHECK_TIMEOUT"
-    while kill -0 "$pid" 2> /dev/null; do
-      sleep "$HEALTH_CHECK_INTERVAL"
-    done
-  else
-    echo "Application didn't start normally. Shutting down!"
-    _stop
-  fi
-else
-  if [[ $COMMANDS == *"$1"* ]]; then
-    exec gosu plone bin/instance "$@"
-  fi
-  exec "$@"
+  exec gosu plone bin/instance console
 fi
+
+# Plone instance helpers
+if [[ $COMMANDS == *"$1"* ]]; then
+  exec gosu plone bin/instance "$@"
+fi
+
+# Custom
+exec "$@"